+++ source/tessellation/EdgeLengthCorrect.cpp
@@ -5,7 +5,7 @@ void CorrectEdgeLength(Tessellation const& tessold,Tessellation const& tessnew,
 {
 	int n=tessold.GetTotalSidesNumber();
 	int npoints=tessold.GetPointNo();
-	lengths.resize(n);
+	lengths.resize((size_t)n);
 	for(int i=0;i<npoints;++i)
 	{
 		vector<int> edgesold=tessold.GetCellEdges(i);
@@ -14,7 +14,7 @@ void CorrectEdgeLength(Tessellation const& tessold,Tessellation const& tessnew,
 		int nedgesnew=(int)edgesnew.size();
 		for(int j=0;j<nedges;++j)
 		{
-			Edge const& edge=tessold.GetEdge(edgesold[j]);
+		  Edge const& edge=tessold.GetEdge(edgesold[(size_t)j]);
 			int n0=edge.neighbors.first;
 			int n1=edge.neighbors.second;
 			bool found=false;
@@ -23,7 +23,7 @@ void CorrectEdgeLength(Tessellation const& tessold,Tessellation const& tessnew,
 				continue;
 			for(int k=0;k<nedgesnew;++k)
 			{
-				Edge const& edgenew=tessnew.GetEdge(edgesnew[(k+j)%nedgesnew]);
+			  Edge const& edgenew=tessnew.GetEdge(edgesnew[(size_t)((k+j)%nedgesnew)]);
 				// are the two edges the same?
 				if(((tessold.GetOriginalIndex(n0)==tessnew.GetOriginalIndex(
 					edgenew.neighbors.first))&&
@@ -33,13 +33,13 @@ void CorrectEdgeLength(Tessellation const& tessold,Tessellation const& tessnew,
 					(tessold.GetOriginalIndex(n1)==tessnew.GetOriginalIndex(
 					edgenew.neighbors.first))))
 				{
-					lengths[edgesold[j]]=0.5*(edge.GetLength()+edgenew.GetLength());
+				  lengths[(size_t)edgesold[(size_t)j]]=0.5*(edge.GetLength()+edgenew.GetLength());
 					found=true;
 					break;
 				}
 			}
 			if(!found)
-				lengths[edgesold[j]]=0.5*edge.GetLength();
+			  lengths[(size_t)edgesold[(size_t)j]]=0.5*edge.GetLength();
 		}
 	}
 }
@@ -49,7 +49,7 @@ void CorrectEdgeLength(Tessellation const& tessold,Tessellation const& tessmid,
 {
 	int n=tessmid.GetTotalSidesNumber();
 	int npoints=tessmid.GetPointNo();
-	lengths.resize(n);
+	lengths.resize((size_t)n);
 	for(int i=0;i<npoints;++i)
 	{
 		vector<int> edgesold=tessold.GetCellEdges(i);
@@ -60,16 +60,16 @@ void CorrectEdgeLength(Tessellation const& tessold,Tessellation const& tessmid,
 		int nedgesmid=(int)edgesmid.size();
 		for(int j=0;j<nedgesmid;++j)
 		{
-			Edge const& edge=tessmid.GetEdge(edgesmid[j]);
+		  Edge const& edge=tessmid.GetEdge(edgesmid[(size_t)j]);
 			int n0=edge.neighbors.first;
 			int n1=edge.neighbors.second;
 			// Don't do double work
 			if((n0<i&&n0>=0)||(n1<i&&n1>=0))
 				continue;
-			lengths[edgesmid[j]]=0.5*edge.GetLength();
+			lengths[(size_t)edgesmid[(size_t)j]]=0.5*edge.GetLength();
 			for(int k=0;k<nedgesnew;++k)
 			{
-				Edge const& edgenew=tessnew.GetEdge(edgesnew[(k+j)%nedgesnew]);
+			  Edge const& edgenew=tessnew.GetEdge(edgesnew[(size_t)((k+j)%nedgesnew)]);
 				// are the two edges the same?
 				if(((tessmid.GetOriginalIndex(n0)==tessnew.GetOriginalIndex(
 					edgenew.neighbors.first))&&
@@ -79,13 +79,13 @@ void CorrectEdgeLength(Tessellation const& tessold,Tessellation const& tessmid,
 					(tessmid.GetOriginalIndex(n1)==tessnew.GetOriginalIndex(
 					edgenew.neighbors.first))))
 				{
-					lengths[edgesmid[j]]+=0.25*edgenew.GetLength();
+				  lengths[(size_t)edgesmid[(size_t)j]]+=0.25*edgenew.GetLength();
 					break;
 				}
 			}
 			for(int k=0;k<nedges;++k)
 			{
-				Edge const& edgeold=tessold.GetEdge(edgesold[(k+j)%nedges]);
+			  Edge const& edgeold=tessold.GetEdge(edgesold[(size_t)((k+j)%nedges)]);
 				// are the two edges the same?
 				if(((tessmid.GetOriginalIndex(n0)==tessold.GetOriginalIndex(
 					edgeold.neighbors.first))&&
@@ -95,7 +95,7 @@ void CorrectEdgeLength(Tessellation const& tessold,Tessellation const& tessmid,
 					(tessmid.GetOriginalIndex(n1)==tessold.GetOriginalIndex(
 					edgeold.neighbors.first))))
 				{
-					lengths[edgesmid[j]]+=0.25*edgeold.GetLength();
+					lengths[(size_t)edgesmid[(size_t)j]]+=0.25*edgeold.GetLength();
 					break;
 				}
 			}
